requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.1.0
idna==3.7
packaging==24.0
psutil==5.9.8
python-dotenv==1.0.1
requests==2.32.2
ruff==0.4.4
urllib3==2.2.1
